\encoding{UTF-8}
\name{sintro-package}
\alias{sintro-package}
\docType{package}
\title{Introduction to R: sintro}
\description{Sample code and supplement to G. Sawitzki: 
Computational Statistics/Introduction to R.}
\details{
\tabular{ll}{
Package: \tab sintro\cr
Type: \tab Package\cr
Version: \tab 0.1-4\cr
Date: \tab 2015-07-18\cr
License: \tab LGPL\cr
}
}
\author{Günther Sawitzki \email{gsawitzki@r-forge.r-project.org}}

\references{
Sawitzki, G.: Computational Statistics/Introduction to R.
	Chapman & Hall/CRC Press, Boca Raton (FL), 2009.
        ISBN: 978-1-4200-8678-2.
	 \url{http://sintro.r-forge.r-project.org/}
}
\keyword{ package }
\keyword{ hplot }
\examples{
#
}
