\encoding{UTF-8}
\name{sintro-package}
\alias{sintro-package}
\docType{package}
\title{sintro: Introduction to R}
\description{Sample code and supplement to G. Sawitzki: 
Computational Statistics/Introduction to R.}
\details{
\tabular{ll}{
Package: \tab sintro\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2012-01-21\cr
License: \tab LGPL\cr
}
}
\author{G. Sawitzki

Maintainer: G. Sawitzki \email{gs@statlab.uni-heidelberg.de}}
\references{
Sawitzki, G.: Computational Statistics/Introduction to R.
	Chapman & Hall/CRC Press, Boca Raton (FL), 2009.
        ISBN: 978-1-4200-8678-2.
	 \url{http://sintro.r-forge.r-project.org/}
}
\keyword{ package }
\keyword{ hplot }
\examples{
#
}
